Uncalibrated pH loop code - no bells and whistles.
This code should, in theory, implement a soft and hard range over which the energia occasionally pumps acid or alkali into the mixture while in the soft range, and immediately pumps acid or alkali when in the hard range.

Note: Need to add goldilocks range between which the energia does nothing. Was part of the initial plan, just forgot to implement it.